requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.18.1
anyio==4.6.2.post1
asyncer==0.0.7
attrs==24.2.0
bidict==0.23.1
certifi==2024.8.30
chainlit==1.3.2
charset-normalizer==3.4.0
chevron==0.14.0
click==8.1.7
dataclasses-json==0.6.7
Deprecated==1.2.15
distro==1.9.0
fastapi==0.115.5
filelock==3.16.1
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.10.0
googleapis-common-protos==1.66.0
grpcio==1.68.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.3
idna==3.10
importlib_metadata==8.5.0
jiter==0.7.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.21
langchain-text-splitters==0.3.2
langsmith==0.1.147
Lazify==0.4.0
literalai==0.0.623
marshmallow==3.23.1
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.54.4
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
orjson==3.10.12
packaging==23.2
propcache==0.2.1
protobuf==5.28.3
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.0
python-dotenv==1.0.1
python-engineio==4.10.1
python-multipart==0.0.9
python-socketio==5.11.4
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
syncer==2.0.3
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.21.0
tomli==2.1.0
tqdm==4.67.0
typing-inspect==0.9.0
typing_extensions==4.12.2
uptrace==1.28.1
urllib3==2.2.3
uvicorn==0.25.0
watchfiles==0.20.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
